+++ src/components/pages/common/download-qr-popup/styled-components.tsx
@@ -1,0 +1,28 @@
+import styled from 'styled-components'
+import { Button } from 'components/common'
+import { Input } from 'linkdrop-ui'
+
+export const InputComponent = styled(Input)`
+  margin-bottom: 10px;
+`
+
+export const PopupFormContent = styled.div`
+  margin-bottom: 20px;
+  padding-top: 20px;
+`
+
+export const PopupForm = styled.form`
+`
+
+
+export const Buttons = styled.div`
+  display: flex;
+  justify-content: end;
+
+  & > button {
+    margin-left: 16px;
+  }
+`
+
+export const WidgetButton = styled(Button)`
+`