+++ src/data/store/index.ts
@@ -1,17 +1,20 @@
 import { createBrowserHistory } from 'history'
 import { combineReducers, createStore, applyMiddleware, compose } from 'redux';
 
-import { userReducer } from './reducers/user/reducer';
+import { userReducer } from './reducers/user/reducer'
 import { contractReducer } from './reducers/contract/reducer'
 import { newRetroDropReducer } from './reducers/campaign/reducer'
 import { campaignsReducer } from './reducers/campaigns/reducer'
 import { qrsReducer } from './reducers/qrs/reducer'
+import { dispensersReducer } from './reducers/dispensers/reducer'
 
-import { UserState } from './reducers/user/types';
-import { ContractState } from './reducers/contract/types';
-import { CampaignState } from './reducers/campaign/types';
-import { CampaignsState } from './reducers/campaigns/types';
-import { QRsState } from './reducers/qrs/types';
+import { UserState } from './reducers/user/types'
+import { ContractState } from './reducers/contract/types'
+import { CampaignState } from './reducers/campaign/types'
+import { CampaignsState } from './reducers/campaigns/types'
+import { QRsState } from './reducers/qrs/types'
+import { DispensersState } from './reducers/dispensers/types';
+
 import thunkMiddleware, { ThunkDispatch, ThunkMiddleware } from "redux-thunk";
 
 import { CampaignActions } from './reducers/campaign/types'
@@ -19,8 +22,9 @@ import { CampaignsActions } from './reducers/campaigns/types'
 import { UserActions } from './reducers/user/types'
 import { ContractActions } from './reducers/contract/types'
 import { QRsActions } from './reducers/qrs/types'
+import { DispensersActions } from './reducers/dispensers/types'
 
-type TActions = CampaignActions & CampaignsActions & UserActions & ContractActions & QRsActions
+type TActions = CampaignActions & CampaignsActions & UserActions & ContractActions & QRsActions & DispensersActions
 export const history = createBrowserHistory()
 
 const reducers = combineReducers({
@@ -28,7 +32,8 @@ const reducers = combineReducers({
   contract: contractReducer,
   campaign: newRetroDropReducer,
   campaigns: campaignsReducer,
-  qrs: qrsReducer
+  qrs: qrsReducer,
+  dispensers: dispensersReducer
 })
 type IAppState = ReturnType<typeof reducers>
 
@@ -44,7 +49,8 @@ export interface RootState {
   contract: ContractState,
   campaign: CampaignState,
   campaigns: CampaignsState,
-  qrs: QRsState
+  qrs: QRsState,
+  dispensers: DispensersState
 }
 
 const store = createStore<RootState, any, any, any>(