+++ src/data/store/reducers/dispensers/async-actions/index.tsx
@@ -1,0 +1,9 @@
+import addDispenser from './add-dispenser'
+import downloadDispenserQR from './download-dispenser-qr'
+import addLinksToQR from './add-links-to-qr'
+
+export {
+  addDispenser,
+  downloadDispenserQR,
+  addLinksToQR
+}