+++ src/data/store/reducers/index.js
@@ -3,13 +3,15 @@ import contract from './contract/reducer'
 import campaign from './campaign/campaign'
 import campaigns from './campaigns/reducer'
 import qrs from './qrs/reducer'
+import dispensers from './dispensers/reducer'
 
 const reducers = {
   userReducer: user,
   contractReducer: contract,
   campaignReducer: campaign,
   campaignsReducer: campaigns,
-  qrsReducer: qrs
+  qrsReducer: qrs,
+  dispensersReducer: dispensers 
 }
 
 export default reducers