+++ src/data/store/reducers/qrs/types.tsx
@@ -1,4 +1,4 @@
-import { ActionType } from 'typesafe-actions';
+import { ActionType } from 'typesafe-actions'
 import * as actions from './actions'
 import { TQRSet } from 'types'
 
@@ -11,4 +11,4 @@ export interface QRsState {
   mappingLoader: number
 }
 
-export type QRsActions = ActionType<typeof actions>;
+export type QRsActions = ActionType<typeof actions>