+++ src/icons/index.js
@@ -29,6 +29,7 @@ import { ReactComponent as PauseIcon } from './pause-icon.svg'
 import { ReactComponent as UnpauseIcon } from './unpause-icon.svg'
 import { ReactComponent as RefundIcon } from './refund-icon.svg'
 import { ReactComponent as ClipboardCopyIcon } from './clipboard-copy-icon.svg'
+import { ReactComponent as ClipboardCopiedIcon } from './clipboard-copied-icon.svg'
 import { ReactComponent as EyeOpenIcon } from './eye-open-icon.svg'
 import { ReactComponent as EyeClosedIcon } from './eye-closed-icon.svg'
 import { ReactComponent as MonitorIcon } from './monitor-icon.svg'
@@ -39,9 +40,12 @@ import { ReactComponent as ContactUsIcon } from './contact-us-icon.svg'
 import { ReactComponent as LegalIcon } from './legal-icon.svg'
 import { ReactComponent as CloseCircleIcon } from './close-circle-icon.svg'
 import { ReactComponent as BookIcon } from './book-icon.svg'
+import { ReactComponent as DispenserIcon } from './dispenser-icon.svg'
 
 const Icons = {
   EthereumLogo,
+  ClipboardCopiedIcon,
+  DispenserIcon,
   GoToOldVersionIcon,
   CloseCircleIcon,
   BookIcon,